refactor: history as module
