+++ studio_app/rout_handlers/__init__.py
@@ -2,5 +2,6 @@ __all__ = [
     'account',
     'book',
     'all_appointments',
-    'appointments'
+    'appointments',
+    'history'
 ]
